remove dynamic typing api
